+++ apps/meteor/client/sidebarv2/Item/Extended.tsx
@@ -1,4 +1,14 @@
-import { Sidebar, IconButton } from '@rocket.chat/fuselage';
+import {
+	SidebarV2Item,
+	SidebarV2ItemAvatarWrapper,
+	SidebarV2ItemCol,
+	SidebarV2ItemRow,
+	SidebarV2ItemTitle,
+	SidebarV2ItemTimestamp,
+	SidebarV2ItemContent,
+	SidebarV2ItemMenu,
+	IconButton,
+} from '@rocket.chat/fuselage';
 import { useEffectEvent, usePrefersReducedMotion } from '@rocket.chat/fuselage-hooks';
 import type { Keys as IconName } from '@rocket.chat/icons';
 import React, { memo, useState } from 'react';
@@ -7,7 +17,7 @@ import { useShortTimeAgo } from '../../hooks/useTimeAgo';
 
 type ExtendedProps = {
 	icon?: IconName;
-	title?: React.ReactNode;
+	title: string;
 	avatar?: React.ReactNode | boolean;
 	actions?: React.ReactNode;
 	href?: string;
@@ -24,7 +34,7 @@ type ExtendedProps = {
 
 const Extended = ({
 	icon,
-	title = '',
+	title,
 	avatar,
 	actions,
 	href,
@@ -37,7 +47,6 @@ const Extended = ({
 	threadUnread: _threadUnread,
 	unread,
 	selected,
-	...props
 }: ExtendedProps) => {
 	const formatDate = useShortTimeAgo();
 	const [menuVisibility, setMenuVisibility] = useState(!!window.DISABLE_ANIMATION);
@@ -47,42 +56,33 @@ const Extended = ({
 	const handleMenu = useEffectEvent((e) => {
 		setMenuVisibility(e.target.offsetWidth > 0 && Boolean(menu));
 	});
-
 	const handleMenuEvent = {
 		[isReduceMotionEnabled ? 'onMouseEnter' : 'onTransitionEnd']: handleMenu,
 	};
 
 	return (
-		<Sidebar.Item selected={selected} highlighted={unread} {...props} {...({ href } as any)} clickable={!!href}>
-			{avatar && <Sidebar.Item.Avatar>{avatar}</Sidebar.Item.Avatar>}
-			<Sidebar.Item.Content>
-				<Sidebar.Item.Content>
-					<Sidebar.Item.Wrapper>
-						{icon}
-						<Sidebar.Item.Title data-qa='sidebar-item-title' className={(unread && 'rcx-sidebar-item--highlighted') as string}>
-							{title}
-						</Sidebar.Item.Title>
-						{time && <Sidebar.Item.Time>{formatDate(time)}</Sidebar.Item.Time>}
-					</Sidebar.Item.Wrapper>
-				</Sidebar.Item.Content>
-				<Sidebar.Item.Content>
-					<Sidebar.Item.Wrapper>
-						<Sidebar.Item.Subtitle className={(unread && 'rcx-sidebar-item--highlighted') as string}>{subtitle}</Sidebar.Item.Subtitle>
-						<Sidebar.Item.Badge>{badges}</Sidebar.Item.Badge>
-						{menu && (
-							<Sidebar.Item.Menu {...handleMenuEvent}>
-								{menuVisibility ? menu() : <IconButton tabIndex={-1} aria-hidden mini rcx-sidebar-item__menu icon='kebab' />}
-							</Sidebar.Item.Menu>
-						)}
-					</Sidebar.Item.Wrapper>
-				</Sidebar.Item.Content>
-			</Sidebar.Item.Content>
-			{actions && (
-				<Sidebar.Item.Container>
-					<Sidebar.Item.Actions>{actions}</Sidebar.Item.Actions>
-				</Sidebar.Item.Container>
-			)}
-		</Sidebar.Item>
+		<SidebarV2Item href={href} selected={selected}>
+			{avatar && <SidebarV2ItemAvatarWrapper>{avatar}</SidebarV2ItemAvatarWrapper>}
+
+			<SidebarV2ItemCol>
+				<SidebarV2ItemRow>
+					{icon && icon}
+					<SidebarV2ItemTitle unread={unread}>{title}</SidebarV2ItemTitle>
+					{time && <SidebarV2ItemTimestamp>{formatDate(time)}</SidebarV2ItemTimestamp>}
+				</SidebarV2ItemRow>
+
+				<SidebarV2ItemRow>
+					<SidebarV2ItemContent unread={unread}>{subtitle}</SidebarV2ItemContent>
+					{badges && badges}
+					{actions && actions}
+					{menu && (
+						<SidebarV2ItemMenu {...handleMenuEvent}>
+							{menuVisibility ? menu() : <IconButton tabIndex={-1} aria-hidden mini rcx-sidebar-v2-item__menu icon='kebab' />}
+						</SidebarV2ItemMenu>
+					)}
+				</SidebarV2ItemRow>
+			</SidebarV2ItemCol>
+		</SidebarV2Item>
 	);
 };